services/dashboard/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
